+++ packages/bsky/tests/data-plane/thread-mutes.test.ts
@@ -1,5 +1,6 @@
 import { RecordRef, SeedClient, TestNetwork, usersSeed } from '@atproto/dev-env'
 import { AtpAgent } from '@atproto/api'
+import { ids } from '../../src/lexicon/lexicons'
 
 describe('thread mutes', () => {
   let network: TestNetwork
@@ -35,7 +36,10 @@ describe('thread mutes', () => {
       { root: rootPost.uriStr },
       {
         encoding: 'application/json',
-        headers: await network.serviceHeaders(alice),
+        headers: await network.serviceHeaders(
+          alice,
+          ids.AppBskyGraphMuteThread,
+        ),
       },
     )
   })
@@ -46,7 +50,7 @@ describe('thread mutes', () => {
         uris: [rootPost.uriStr, replyPost.uriStr],
       },
       {
-        headers: await network.serviceHeaders(alice),
+        headers: await network.serviceHeaders(alice, ids.AppBskyFeedGetPosts),
       },
     )
     expect(res.data.posts[0].viewer?.threadMuted).toBe(true)
@@ -60,7 +64,12 @@ describe('thread mutes', () => {
 
     const notifsRes = await agent.api.app.bsky.notification.listNotifications(
       {},
-      { headers: await network.serviceHeaders(alice) },
+      {
+        headers: await network.serviceHeaders(
+          alice,
+          ids.AppBskyNotificationListNotifications,
+        ),
+      },
     )
     expect(notifsRes.data.notifications.length).toBe(0)
   })
@@ -72,7 +81,12 @@ describe('thread mutes', () => {
 
     const notifsRes = await agent.api.app.bsky.notification.listNotifications(
       {},
-      { headers: await network.serviceHeaders(alice) },
+      {
+        headers: await network.serviceHeaders(
+          alice,
+          ids.AppBskyNotificationListNotifications,
+        ),
+      },
     )
     expect(notifsRes.data.notifications.length).toBe(0)
   })
@@ -84,7 +98,12 @@ describe('thread mutes', () => {
 
     const notifsRes = await agent.api.app.bsky.notification.listNotifications(
       {},
-      { headers: await network.serviceHeaders(alice) },
+      {
+        headers: await network.serviceHeaders(
+          alice,
+          ids.AppBskyNotificationListNotifications,
+        ),
+      },
     )
     expect(notifsRes.data.notifications.length).toBe(0)
   })
@@ -96,7 +115,12 @@ describe('thread mutes', () => {
 
     const notifsRes = await agent.api.app.bsky.notification.listNotifications(
       {},
-      { headers: await network.serviceHeaders(alice) },
+      {
+        headers: await network.serviceHeaders(
+          alice,
+          ids.AppBskyNotificationListNotifications,
+        ),
+      },
     )
     expect(notifsRes.data.notifications.length).toBe(0)
   })
@@ -106,7 +130,10 @@ describe('thread mutes', () => {
       { root: rootPost.uriStr },
       {
         encoding: 'application/json',
-        headers: await network.serviceHeaders(alice),
+        headers: await network.serviceHeaders(
+          alice,
+          ids.AppBskyGraphUnmuteThread,
+        ),
       },
     )
   })
@@ -117,7 +144,7 @@ describe('thread mutes', () => {
         uris: [rootPost.uriStr, replyPost.uriStr],
       },
       {
-        headers: await network.serviceHeaders(alice),
+        headers: await network.serviceHeaders(alice, ids.AppBskyFeedGetPosts),
       },
     )
     expect(res.data.posts[0].viewer?.threadMuted).toBe(false)
@@ -133,7 +160,12 @@ describe('thread mutes', () => {
 
     const notifsRes = await agent.api.app.bsky.notification.listNotifications(
       {},
-      { headers: await network.serviceHeaders(alice) },
+      {
+        headers: await network.serviceHeaders(
+          alice,
+          ids.AppBskyNotificationListNotifications,
+        ),
+      },
     )
     expect(notifsRes.data.notifications.length).toBe(4)
   })